requirements.txt
absl-py==1.4.0
aiohttp==3.8.4
aiohttp-cors==0.7.0
aioice==0.7.6
aiortc==1.4.0
aiosignal==1.3.1
asgiref==3.6.0
astor==0.8.1
async-timeout==4.0.2
asynctest==0.13.0
attrs==22.2.0
av==10.0.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
cmake==3.25.2
colorama==0.4.6
cryptography==39.0.2
Django==3.2.18
django-cors-headers==3.14.0
django-sslserver==0.22
djangorestframework==3.14.0
dlib==19.20.0
dnspython==2.3.0
face-recognition==1.3.0
face-recognition-models==0.3.0
filetype==1.2.0
Flask==2.2.4
Flask-Cors==3.0.10
frozenlist==1.3.3
gast==0.2.2
google-crc32c==1.5.0
google-pasta==0.2.0
grpcio==1.30.0
h5py==2.10.0
idna==3.4
importlib-metadata==6.6.0
imutils==0.5.3
itsdangerous==2.1.2
Jinja2==3.1.2
Keras==2.3.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
Markdown==3.4.3
MarkupSafe==2.1.2
multidict==6.0.4
netifaces==0.11.0
numpy==1.19.0
opencv-python==4.3.0.36
opt-einsum==3.2.1
packaging==23.1
pdf2image==1.16.3
Pillow==9.5.0
protobuf==3.12.2
psycopg2==2.9.6
pycparser==2.21
pyee==9.0.4
pylibsrtp==0.8.0
pyOpenSSL==23.0.0
pytesseract==0.3.10
pytz==2022.7.1
PyYAML==5.3.1
requests==2.29.0
scipy==1.5.1
six==1.15.0
sqlparse==0.4.3
tensorboard==1.15.0
tensorflow==1.15.2
tensorflow-estimator==1.15.1
termcolor==1.1.0
typing_extensions==4.5.0
urllib3==1.26.15
Werkzeug==2.2.3
wrapt==1.12.1
yarl==1.8.2
zipp==3.1.0
